pyproject.toml
[project]
name = "pytest-approval"
version = "0.8.0"
description = "A simple approval test library utilizing external diff programs such as PyCharm and Visual Studio Code to compare approved and received output."
readme = "README.md"
license = "LGPL-3.0-only"
authors = [{email="ohsome@heigit.org"}]
classifiers = ["Framework :: Pytest"]
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.3.5",
]

[dependency-groups]
dev = [
    "pytest-subprocess>=1.5.3",
    "ruff>=0.11.10",
]

[project.optional-dependencies]
image = [
    "numpy>=2.3.1",
    "pillow>=11.3.0"
]

[project.entry-points.pytest11]
pytest_approval = "pytest_approval.plugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.build-backend]
module-root = ""

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle Error
    "F",    # Pyflakes
    "I",    # isort
    "N",    # pep8-nameing
    "Q",    # flake8-quotes
    "W",    # pycodestyle Warning
    "C90",  # mccabe
    "TRY",  # tryceratops
    "RUF",  # ruff-specific rules
    "FURB", # refurb
    "PERF", # perflint
    "S",    # bandit
    "B",    # bugbear
    "SIM",  # simplify
]

[tool.ruff.lint.extend-per-file-ignores]
"tests/*" = ["S101"]  # S101 usage of asserts
